myproj/requirements/dev.txt
-r common.txt

django-debug-toolbar==4.3.0
